+++ packages/cached/lib/cached.dart
@@ -12,5 +12,10 @@ Builder cachedBuilder(BuilderOptions options) {
       ),
     ],
     '.cached.dart',
+    header: '''
+    // coverage:ignore-file
+    // GENERATED CODE - DO NOT MODIFY BY HAND
+    // ignore_for_file: type=lint
+    ''',
   );
 }