+++ packages/cached_annotation/lib/cached_annotation.dart
@@ -1,7 +1,9 @@
 library cached_annotation;
 
+export 'src/cache_key.dart';
 export 'src/cached.dart';
 export 'src/clear_all_cached.dart';
 export 'src/clear_cached.dart';
+export 'src/ignore.dart';
 export 'src/ignore_cache.dart';
 export 'src/with_cache.dart';