+++ packages/cached_annotation/CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
+## 1.1.0
+* Add `@IgnoreCache()` annotation
+* Add `@CacheKey()` annotation
 
+## 1.0.0
 * Initial version of cached_annotation library